--- stories/2 - Presets/Tree/components/Handle/Handle.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-
-import styles from './Handle.module.css';
-
-interface Props extends React.HTMLAttributes<HTMLDivElement> {}
-
-export function Handle(props: Props) {
-  return (
-    <div
-      className={styles.Handle}
-      tabIndex={0}
-      data-cypress="draggable-handle"
-      {...props}
-    >
-      <svg viewBox="0 0 20 20">
-        <path d="M7 2a2 2 0 1 0 .001 4.001A2 2 0 0 0 7 2zm0 6a2 2 0 1 0 .001 4.001A2 2 0 0 0 7 8zm0 6a2 2 0 1 0 .001 4.001A2 2 0 0 0 7 14zm6-8a2 2 0 1 0-.001-4.001A2 2 0 0 0 13 6zm0 2a2 2 0 1 0 .001 4.001A2 2 0 0 0 13 8zm0 6a2 2 0 1 0 .001 4.001A2 2 0 0 0 13 14z"></path>
-      </svg>
-    </div>
-  );
-}
