--- stories/2 - Presets/Tree/components/Handle/index.ts
@@ -1,1 +1,0 @@
-export {Handle} from './Handle';
