+++ stories/2 - Presets/Tree/components/TreeItem/TreeItem.module.css
@@ -10,16 +10,12 @@
     padding: 5px;
 
     .TreeItem {
-      --vertical-padding: 15px;
+      --vertical-padding: 5px;
 
-      min-width: 100px;
-      border-radius: 3px;
+      padding-right: 24px;
+      border-radius: 4px;
       box-shadow: 0px 15px 15px 0 rgba(34, 33, 81, 0.1);
     }
-  }
-
-  &:not(.clone) {
-    animation: fadeIn 450ms ease;
   }
 
   &.ghost {
@@ -28,21 +24,31 @@
     .TreeItem {
       > * {
         box-shadow: none;
+        background-color: transparent;
       }
     }
   }
 }
 
 .TreeItem {
-  --vertical-padding: 20px;
+  --vertical-padding: 10px;
 
   position: relative;
   display: flex;
-  padding: var(--vertical-padding) 20px;
+  align-items: center;
+  padding: var(--vertical-padding) 10px;
   background-color: #fff;
   border: 1px solid #dedede;
   color: #222;
   box-sizing: border-box;
+}
+
+.Text {
+  flex-grow: 1;
+  padding-left: 0.5rem;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  overflow: hidden;
 }
 
 .Count {
@@ -59,13 +65,4 @@
   font-size: 0.8rem;
   font-weight: 600;
   color: #fff;
-}
-
-@keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
 }
