+++ stories/2 - Presets/Tree/components/index.ts
@@ -1,2 +1,1 @@
-export {Handle} from './Handle';
 export {TreeItem, SortableTreeItem} from './TreeItem';
