+++ stories/components/Item/index.ts
@@ -1,1 +1,2 @@
 export {Item} from './Item';
+export {Handle, Remove} from './components';
