+++ stories/components/index.ts
@@ -2,7 +2,7 @@ export {Button} from './Button';
 export {ConfirmModal} from './ConfirmModal';
 export {Axis, Draggable} from './Draggable';
 export {Droppable} from './Droppable';
-export {Item} from './Item';
+export {Item, Handle, Remove} from './Item';
 export {FloatingControls} from './FloatingControls';
 export {Grid} from './Grid';
 export {GridContainer} from './GridContainer';
